api/requirements.txt
flask
flask-session
flask-limiter
redis
openai
azure-storage-blob
PyPDF2
python-docx
faiss-cpu
